Actions: improve field handling in actions
The Field API in QS has one issue: it does not report any status when an
invalid field (by name) is requested. All Promises remain silent, while
they do resolve when concerning a valid field.

To counter this, the getField() method helps to report on invalid field
requests. Such situations may occur when the field name's expression is
wrongly saved as a Field reference instead of a string. Further
explanation will be included in the documentation.